requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.0
folium==0.20.0
gunicorn==21.2.0
python-dotenv==1.0.0
reportlab==4.0.7
weasyprint==60.2
Pillow==10.1.0

